+++ aider/website/HISTORY.md
@@ -16,14 +16,15 @@ cog.out(text)
 
 # Release history
 
-### main branch
+### Aider v0.52.0
 
-- Aider now offers to run shell commands related to changes it makes:
+- Aider now offers to run shell commands:
   - Launch a browser to view updated html/css/js.
   - Install new dependencies.
   - Run DB migrations. 
   - Run the program to exercise changes.
   - Run new test cases.
+- `/read` and `/drop` now expand `~` to the home dir.
 - Show the active chat mode at aider prompt.
 - New `/reset` command to `/drop` files and `/clear` chat history.
 - New `--map-multiplier-no-files` to control repo map size multiplier when no files are in the chat.
@@ -32,7 +33,7 @@ cog.out(text)
 - Improved formatting of token reports and confirmation dialogs.
 - Default OpenAI model is now `gpt-4o-2024-08-06`.
 - Bumped dependencies to pickup litellm bugfixes.
-- Aider wrote 64% of the code in this release.
+- Aider wrote 68% of the code in this release.
 
 ### Aider v0.51.0